Fixing security

+++ src/Controller/PaymentController.php
@@ -9,20 +9,32 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Security;
 
 /**
  * @Route("/payment")
  */
 class PaymentController extends AbstractController
 {
+    private $security;
+
+    public function __construct(Security $security)
+    {
+        $this->security = $security;
+    }
+
     /**
      * @Route("/", name="payment_index", methods={"GET"})
      */
     public function index(PaymentRepository $paymentRepository): Response
     {
-        return $this->render('payment/index.html.twig', [
-            'payments' => $paymentRepository->findAll(),
-        ]);
+        if ($this->security->isGranted('ROLE_ADMIN')){
+            return $this->render('payment/index.html.twig', [
+                'payments' => $paymentRepository->findAll(),
+            ]);
+        } else {
+            return $this->redirectToRoute('default');
+        }
     }
 
     /**
@@ -30,22 +42,26 @@ class PaymentController extends AbstractController
      */
     public function new(Request $request): Response
     {
-        $payment = new Payment();
-        $form = $this->createForm(PaymentType::class, $payment);
-        $form->handleRequest($request);
+        if ($this->security->isGranted('ROLE_ADMIN')){
+            $payment = new Payment();
+            $form = $this->createForm(PaymentType::class, $payment);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->persist($payment);
-            $entityManager->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->persist($payment);
+                $entityManager->flush();
 
-            return $this->redirectToRoute('payment_index');
+                return $this->redirectToRoute('payment_index');
+            }
+
+            return $this->render('payment/new.html.twig', [
+                'payment' => $payment,
+                'form' => $form->createView(),
+            ]);
+        } else {
+            return $this->redirectToRoute('default');
         }
-
-        return $this->render('payment/new.html.twig', [
-            'payment' => $payment,
-            'form' => $form->createView(),
-        ]);
     }
 
     /**
@@ -53,9 +69,13 @@ class PaymentController extends AbstractController
      */
     public function show(Payment $payment): Response
     {
-        return $this->render('payment/show.html.twig', [
-            'payment' => $payment,
-        ]);
+        if ($this->security->isGranted('ROLE_ADMIN')){
+            return $this->render('payment/show.html.twig', [
+                'payment' => $payment,
+            ]);
+        } else {
+            return $this->redirectToRoute('default');
+        }
     }
 
     /**
@@ -63,21 +83,25 @@ class PaymentController extends AbstractController
      */
     public function edit(Request $request, Payment $payment): Response
     {
-        $form = $this->createForm(PaymentType::class, $payment);
-        $form->handleRequest($request);
+        if ($this->security->isGranted('ROLE_ADMIN')){
+            $form = $this->createForm(PaymentType::class, $payment);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $this->getDoctrine()->getManager()->flush();
 
-            return $this->redirectToRoute('payment_index', [
-                'id' => $payment->getId(),
+                return $this->redirectToRoute('payment_index', [
+                    'id' => $payment->getId(),
+                ]);
+            }
+
+            return $this->render('payment/edit.html.twig', [
+                'payment' => $payment,
+                'form' => $form->createView(),
             ]);
+        } else {
+            return $this->redirectToRoute('default');
         }
-
-        return $this->render('payment/edit.html.twig', [
-            'payment' => $payment,
-            'form' => $form->createView(),
-        ]);
     }
 
     /**
@@ -85,12 +109,16 @@ class PaymentController extends AbstractController
      */
     public function delete(Request $request, Payment $payment): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$payment->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($payment);
-            $entityManager->flush();
-        }
+        if ($this->security->isGranted('ROLE_ADMIN')){
+            if ($this->isCsrfTokenValid('delete'.$payment->getId(), $request->request->get('_token'))) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($payment);
+                $entityManager->flush();
+            }
 
         return $this->redirectToRoute('payment_index');
+        } else {
+            return $this->redirectToRoute('default');
+        }
     }
 }
